+++ Generals/Code/GameEngine/Source/GameClient/GUI/GUICallbacks/ControlBarCallback.cpp
@@ -503,7 +503,7 @@ void ShowControlBar( Bool immediate )
 
 	TheControlBar->showSpecialPowerShortcut();
 
-	Int id = (Int)TheNameKeyGenerator->nameToKey(AsciiString("ControlBar.wnd:ControlBarParent"));
+	Int id = (Int)(uintptr_t)TheNameKeyGenerator->nameToKey(AsciiString("ControlBar.wnd:ControlBarParent"));
 	GameWindow *window = TheWindowManager->winGetWindowFromId(NULL, id);
 
 	if (window)
@@ -538,7 +538,7 @@ void HideControlBar( Bool immediate )
 
 	TheControlBar->hideSpecialPowerShortcut();
 
-	Int id = (Int)TheNameKeyGenerator->nameToKey(AsciiString("ControlBar.wnd:ControlBarParent"));
+	Int id = (Int)(uintptr_t)TheNameKeyGenerator->nameToKey(AsciiString("ControlBar.wnd:ControlBarParent"));
 	GameWindow *window = TheWindowManager->winGetWindowFromId(NULL, id);
 
 	if (window)
@@ -574,7 +574,7 @@ void ToggleControlBar( Bool immediate )
 
 	toggleReplayControls();
 
-	Int id = (Int)TheNameKeyGenerator->nameToKey(AsciiString("ControlBar.wnd:ControlBarParent"));
+	Int id = (Int)(uintptr_t)TheNameKeyGenerator->nameToKey(AsciiString("ControlBar.wnd:ControlBarParent"));
 	GameWindow *window = TheWindowManager->winGetWindowFromId(NULL, id);
 
 	if (window)
