+++ Generals/Code/GameEngine/Source/GameClient/GUI/GameWindowManager.cpp
@@ -754,7 +754,7 @@ Int GameWindowManager::winSetFocus( GameWindow *window )
 	if( (m_keyboardFocus) && (m_keyboardFocus != window) )
 	{
 		Bool wf;	// dummy var, ignored, but must be passed
-		winSendSystemMsg( m_keyboardFocus, GWM_INPUT_FOCUS, FALSE, (WindowMsgData)&wf );
+		winSendSystemMsg( m_keyboardFocus, GWM_INPUT_FOCUS, FALSE, (WindowMsgData)(uintptr_t)&wf );
 	}
 
 	// Set focus to new window
@@ -766,7 +766,7 @@ Int GameWindowManager::winSetFocus( GameWindow *window )
 
 		for (;;)
 		{
-			winSendSystemMsg( window, GWM_INPUT_FOCUS, TRUE, (WindowMsgData)&wantsFocus );
+			winSendSystemMsg( window, GWM_INPUT_FOCUS, TRUE, (WindowMsgData)(uintptr_t)&wantsFocus );
 			if (wantsFocus)
 				break;
 
@@ -971,10 +971,10 @@ WinInputReturnCode GameWindowManager::winProcessMouseEvent( GameWindowMessage ms
 
 						newRegion.hi.x = newRegion.lo.x + grabSize.x;
 						newRegion.hi.y = newRegion.lo.y + grabSize.y;
-						if( newRegion.hi.x > (Int)TheDisplay->getWidth() )
-							newRegion.hi.x = (Int)TheDisplay->getWidth();
-						if( newRegion.hi.y > (Int)TheDisplay->getHeight() )
-							newRegion.hi.y = (Int)TheDisplay->getHeight();
+						if( newRegion.hi.x > (Int)(uintptr_t)TheDisplay->getWidth() )
+							newRegion.hi.x = (Int)(uintptr_t)TheDisplay->getWidth();
+						if( newRegion.hi.y > (Int)(uintptr_t)TheDisplay->getHeight() )
+							newRegion.hi.y = (Int)(uintptr_t)TheDisplay->getHeight();
 
 						newRegion.lo.x = newRegion.hi.x - grabSize.x;
 						newRegion.lo.y = newRegion.hi.y - grabSize.y;
